Don't hang if the I2C lines are shorted together.
My keyboard has a TRRS jack with integrated switch that shorts SDA and
SCL together if there's nothing plugged in. The TWI hardware never
signals completion if this is the case, so the keyboard hangs.

This change allows the "master" half of the keyboard to work even in
case of such a major I2C fault.